DSP-17444: Fix UnsafeRowReader and Writer Conversions
Previously this code was incorrect in it's hadnling of timestamps, udts,
and a variety of other types. To fix this we now perform a two stage
conversion, first converting to a SparkSQL Scala understandable type, then
we use the built in Catalyst converters to perform a subsequent
conversion from the Scala type to a Catalyst compatible type. This is
important because several Scala Types have slight modifications before
they become valid in the Catalyst Internal representation. For example,
all timestamps are multiplied by 1000 before we being processed. By
using the Catlayst converter we can take account of all these
modifications.