SPARKC-558: RegularStatements not Cached by SessionProxy
Previously the case matching in the SessionProxy incorrectly stated that
the RegularStatement class was a String, this meant that Regular
Statements would never actually get matched. To fix this we correct the
class type and in addition we add some tests to actually check the
cache. This still can't avoid all reprepares since the TrieMap is a "at
least once" guarentee and can generate extra calls sometimes.