DSP-19063: DSE DirectJoin Output Check too Strict
There are certain joins whose end projections will end up different after a direct join application

For example

    Given TableOne (A, B, C) PK (A,B)
    Given TableTwo (A, B)

    Join TableOne and TableTwo ON A = A WHERE C = 5

    The original plan will look like

    Join [ A = A] Output (A)
           |
      ++++++++++++++++++++++++++++++++++++++++++++++++++++
      |                                                  |
    Filter (C = 5)  (A ,C)                  TableTwo Read (A)
      |
    TableOne Read (A, C)
    Will be organized into

    Filter (C =5) (A, C)
          |
        DirectJoin [ A = TableOne A] (A,C)
          |
     TableTwo Read(A) (A)

Both of these are valid but there is a very strict check on the output of the new plan matching
the output of the old plan. What's really important is that the output plan includes all of the
old plan outputs, but it's fine if it adds additional ones.