DSP-18681: Modify TTL and WriteTime to Work with Projections
Previously the the TTL and WriteTime functions would break if they were
drawn from a projection rather than from the original source relation.
Something like .select("v").select("TTL(v)"). This is because the
original projection would not have the hand replaced ttl attribute we
add during the resolution phase. We fix this by adding that missing
attribute to all children of a node where we add the new TTL attribute
reference.

Once we have placed the new CassandraTTL column in the underlying
Relation we still need to make sure it is surfaced up to the original
function call location. To do this, we look for function calls that are
missing their input column. For the children of these nodes we go to the
leaves and step back up the tree one node at a time. If the node has the
real column as part of its input then we add it to the output. If the
node has no real TTL column as part of it's input, then there is no
valid TTL and we return a null.

This is explicitly for cases where you union a C* table with a Dataframe
so that while a column is "real" on one side of the tree, it is a null
or fake column on the other side.

DSP-18681: Refactor application of TTL Resolution Rule

Previously as a post-hoc resolution rule we attempted to leave the
logical tree in a fully resolved state. This proved to be very difficult
to make work when other columns are merged into unions. To get around
this we instead change the rule to a normal Resolution rule which allows
us to leave elements of the tree unresolved for a second pass.

With this change we are able to look for places where we have added an
unresolved ttl reference, and more unresolved ttl references, and only
add resolved attribute references to leaves.

DSP-18681: Prevent Union Nullability Checks Failing

The union outputset resolves the nullability of it's attributes whenever
you ask it what it's output is. This is a problem because unresolved
attributes do not have nullability. We actually do know that our
records are nullable so we make a fake placeholder class for our
unresolved element which is nullable. This allows us to ignore Union's
issue.

DSP-19403: Handle case insensitive ttl and writetime options
SpaekSQL data source options are case insesitive. THe code is fixed to properly
search C* columns igonorring case.