DSP-18857: Fix CustomTableScanMethodSpec
When testing against DSE we need to use a DseCassandraConnectionFactory,
otherwise the metadata will not correctly be updated and commands will
fail. If we need to support pure OSS and this becomes an issue in the
future we should have two versions of the DummyConnector object, one
which uses CqlSession and one that uses DseSession. For the moment I
don't think the complexity warrents it.