DSP-18857: CassandraRDD Passing
Had to fix the string format for the dates passed into the converter.

Attempted to fix the Annotation based type mapping but the new
Driver annotation classes remove all annotation info from runtime. Even
with the generated classes it becomes impossible to determine how a
particular field was annotated. For now we will ignore the test but I
did fix the build so we can actually build annotated classes.