DSP-18857 fix developer mode for CassandraSSLConnectorSpec
Dev mode was broken for auth/ssl clusters because it relies on CcmConfig
string representation (it uses CcmConfig.toString() to verify if a
cluster for this particular config string already exists). Auth and Ssl
clusters used to contain temporal paths in config which made
CcmConfig.toString different for each execution.

Temporal paths are substituted with constant paths to /tmp/ directory.