DSP-18857: Remove Custom Routing Key Generator Code
The Driver in 4.0 now automatically sets RoutingKeys on all bound
statements which set partition keys. For us this is all requests, we
never set partition keys as strings in the statements. So we can safely
remove all code that sets routing keys in the SCC. This also simplifies
the routing key generator and batch statement builder.

The new rules are

If a partition key is bound, the key can be generated
If BatchStatment is used, the first available child with a routing key
is used