Adds a MultiplexingSchemaListener
This listener allows users to add or remove new SchemaListeners at
runtime. Listeners can be added and removed to the Multiplexing listener
which calls all added listeners in order.

Backed by a TrieMap to allow for multi-threaded usage. Doesn't block on
adding or removing listeners but events triggered during the additions
may not be applied to newly added listeners.