SPARKC-458: Fix for 2.12.10 Build Issues and Review
The Spark Packages plugin was automatically adding a default spark
version of 1.4.0 to all projects which didn't inherit the
sparkPackagesSettings group. With this setting added to the
cassandra-embedded project the Scala 2.12 integration test can compile.
Additionally the SparkRepl class exists in Scala Specific source
directories, there was no Scala 2.12 version so that is added as well.

The Cassandra Java Util test also required a small fix because it was
doing a strict string comparison of typetag's string representations. In
Scala 2.12 java.lang is no longer added to Java types (getSimpleName) vs
(getName) causing a bunch of mismatches. We remove all "java.lang." from
the strings before comparing to fix the mismatch and be compatible with
both Scala versions.

In Scala 2.12 there seems to be a sensitivity to untyped implicits and their
position in the code. Previously, the unannotated implicit connector in
the CassandraSourceRelation was declared after it's usage in
"saveToCassandra", in Scala 2.12 this caused the implicit to be ignored
and the default of (CassandraConnector(sparkConf)) to be used instead.
This broke configuring the DataSource using an alternative connection
conf. In Scala 2.11 there is no problem with the ordering.

To fix this we move and explicitly type the implicit val declarations to the
top of the file.

Adds travis configuration for running 2.12 Integration Tests