SPARKC-573: Enable Travis CI
Mostly this is straightforward. We need to change a few of the
environment variables to enable CCM to work properly. On the test
side several tests needed to be adjusted so that they would be a
little less stressful on travis's smaller nodes. Mostly this invloved
using the AsyncExecutor class which handles retries and throttling much
better than just firing off a future and hoping.