SPARKC-589: DataSource V2 Support
* Namespace Support

Namespaces in DatasourceV2 Catalogs correspond to Keyspaces in
Cassandra. Create, Delete, Alter, List are all supported.

Available properties to set are class, durable_writes,
replication_factor. All other properties are assumed to be DC's in a
network topology strategy.

Example Create Statement

CREATE DATABASE IF NOT EXISTS someks WITH DBPROPERTIES (class='SimpleStrategy',replication_factor='5')

* TableCatalog Support

Adds the TableSupport to the CassandraCatalog Class. We support almost
all things that are valid in CQL.

Table Creation: Tables can be created using standard Spark SQL syntax.
We allow the use of Partitioned By to set Partitioning keys as well as
setting the partition_key parameter.

Currently there is no native terms in SparkSQL for clustering or sorting
so that must currently be set as a parameter, clustering_key but in the
future we should be able to use a native term.

All other properties can be provided either as a String or a Map, Maps
must be formatted like a cassandraOption='{key=value, key=value,
key=value'}.

Example Create Statement:

CREATE TABLE ks.table (
key_1 Int, key_2 Int, key_3 Int,
cc1 STRING, cc2 String, cc3 String,
value String)
PARTITIONED BY (key_1, key_2, key_3)
TBLPROPERTIES (clustering_key='cc1.asc, cc2.desc, cc3.asc')

List Tables:
Listing tables within a keyspace

Load Table:
Although the table implementation does not yet support reading and
writing

AlterTable: Usual cassandra restrictions apply,
  It is forbidden to drop Columns which are part of the primary key
  Table properties can only be set
  Only normal columns can be added or removed

DropTable: All tables can be dropped

RenameTable: Forbidden

* Remove Spark 2.11 as a build Target

The move to Spark 3.0.0 removes 2.11 Compatibility

* Fix up translation of UTF8 and Tuple Types

Previously we were allowed to use Product Types and UTF8 String when
building internal RDD representations for Dataframes. They are no longer
accepted so they have been replaced with equilvelent representations.

* DataSourceV2 Read Support

In this commit we refactor out a large deal of the CassandraTableScanRDD
code into pure functions held in ScanHelper. This allows us to reuse
this code in the new DSV2 Read path which no longer allows directly
delegating to an RDD. In DSV2 a new RDD is constructed using
ReaderFactories which generatate their own readers. We have also added
basic code around passing C* Clustering information which allows for
Spark to take advantage of underlying C* partitioning during operations.

* DSV2 Basic Write Support

Like with reads, the first step was to refactor out the code from
Tablewriter which assumed that an iterator was provider. A new class
AysncDriverWriter was created to handle rows one at a time as they are
fed into the writer. These can come from a iterator but this allows us
to fit the same code path into the DSV2 Api as well.

* Count Support

Previously we split responsibilites between ScanBuilder and Scan for
doing some of our analysis of the query to be built. I realized it makes
more logical sense to move this all as high as possible in the call
stack. This makes it easier to apply count optimizations without passing
through lots of "magic" empty arrays.

After doing so Count is now supported by a second PartitionReaderFactory
which acts like the Normal Scan Reader but instead of generating 1 for 1
rows out of Cassandra, it preforms a count, then generates empty rows
for the count result.

* Added Naming suggestions to NoSuchTable and NoSuchKeyspace

Unfortunately these are only exercised for certain SparkSql verbs. Many
operations will not use the custom extension messages. Hopefully in a
future version of Spark we can support all usecases.

* TTL/Writetime Metadata Functions

Only a slight refactoring was required to make the Metadata functions
work again. Basically all references to "relation" are changed to
DSV2Relation and we can modify the table there.

One limitation that is introduced is that the option process of
specifying ttls "ttl.colname = ttlName" is no longer valid. Because of
some of the new verification for output, we need to modify the plan
before the scan to have the correct number of ouputs. This means that
only via our Rule can we handle new columns. All behaviors with the
functions are as before so you can still select ("ttl(col)").

* Fix ScalaDoc Generation

Due to an issue with how ScalaDoc in Scala 2.12 attempts to generate
docs for Java Classes, the doc portion of publishing was broken. The fix
used by Spark (and suggested on the Scala Bug) is to simply turn off the
Java based files during doc generation.

See [Spark-22855] https://github.com/apache/spark/pull/20042/files
And [Scala Bug #11365] https://github.com/scala/bug/issues/11365

* DirectJoin Support for DSV2

The main changes here take place inside of the DirectJoinStrategy. Since
we no longer are looing for CassandraSource relations we need to swap
all references for DatasourceV2ScanRelations and other analgous classes.
We also no longer have an RDD which we can pass through to the
DirectJoinExec.

One Key issue is that AttributeReferences now contain possible metdata
strings based on their origin. This makes doing "contains" or equals on
Attribute references now extremely unreliable. To account for this we've
switched all Attribute matching to ExprId matching. This matches
elements based on their global plan identifier rather, than if they have
identical metdata.

* InClauseJoin Optimization

Much like the Writer and Reader, this required extracting out a lot of
code from the CassandraJoinRDD into JoinHelper. This way we can access
the same functions outside of an RDD context in the new InJoinReader.
This new class basically mimics the "compute" method of a CassandraJoin
but uses the iterators from InClauses as keys.

The algorithm for doing the cross product has slightly changed as we no
longer can have spark's "cross-product" do the cross for us since we are
no longer parallelizing the in clause values. Instead we parallelize the
in clauses in the reader, then each reader uses the PartitionID and
Total number of partitions to choose which records from a crossproduct
for a particular reader to execute.