SPARKC-590: Java 11 Support
* Ambiguity in ByteBuffer Api

Java 11 introduces another method named position which takes a single
arg, scala couldn't tell which one to use without the paren's showing we
want the 0 arg version of the function call.

* Fix TimestampParser Underspecification

From the Docs

> Offset X and x: This formats the offset based on the number of pattern letters. One letter outputs just the hour, such as '+01', unless the minute is non-zero in which case the minute is also output, such as '+0130'. Two letters outputs the hour and minute, without a colon, such as '+0130'. Three letters outputs the hour and minute, with a colon, such as '+01:30'. Four letters outputs the hour and minute and optional second, without a colon, such as '+013015'. Five letters outputs the hour and minute and optional second, with a colon, such as '+01:30:15'. Six or more letters throws IllegalArgumentException. Pattern letter 'X' (upper case) will output 'Z' when the offset to be output would be zero, whereas pattern letter 'x' (lower case) will output '+00', '+0000', or '+00:00

So we need up to 5 X's to satisify the test cases we run, in Java 8 It
simply ignored the non parsed Characters. In Java 11 we need to
explicitly account for all characters in the timestamp string.

* Remove Repl Spec

The repl spec has always been flakey and it has once again broken with
this Java upgrade. Since the things it tested are all pretty much not
recommended now, and the test is manily a burden, I removed it.