SPARKC-621 move timeuuid non-equal predicates check at the end
Before this commit the check was executed in
BasicCassandraPredicatePushDown at the beginning of the push down
evaluation. This interrupted push down evaluation without giving
other predicates rules a chance to push down the predicate.
Since thanks to SAI we may now push timeuuid predicates, this
commit moves the check at the very end of the push down evaluation.