SPARKC-642 fix TwoNodeCluster integration test setup
ReplicaRepartitionedCassandraRDDSpec based on TwoNodeCluster was not
working for DSE 6.8 as this db version relies on metadata dir.
Before this change the metadata dir was shared among
all the nodes in the cluster (which caused Gossip failures that
eventually lead to test failure).
After this change all the nodes in the cluster use separate
metadata dirs.
Additionally, autoboostrapping is now enabled as all the nodes in
TwoNodeCluster need to have data from relevant parts of the token
ring.